fix(api): move shared helpers out of api/_lib to lib/ at repo root
Vercel serverless functions fail to load with FUNCTION_INVOCATION_FAILED
when they import from api/_lib/. Underscore-prefixed folders inside
api/ are excluded from the deployed function bundle, so every v1/*
endpoint crashed at module load before the handler could run.

Verified by adding two diagnostic endpoints:
- api/v1/diag.ts (no imports) → 200 OK
- api/v1/diag-lib.ts (imports from _lib) → FUNCTION_INVOCATION_FAILED

Moving the helpers out of api/ altogether avoids the underscore
convention entirely. File-trace at build time still follows the
imports, so the helpers ship in each function bundle as expected.
