--- api/v1/diag.ts
@@ -1,5 +1,0 @@
-import type { VercelRequest, VercelResponse } from '@vercel/node'
-
-export default function handler(_req: VercelRequest, res: VercelResponse) {
-  res.status(200).json({ ok: true, step: 'no-imports' })
-}
