+++ api/v1/feedback-shares/[shareId]/prompt.ts
@@ -1,9 +1,9 @@
 import type { VercelRequest, VercelResponse } from '@vercel/node'
-import { requireReviewer } from '../../../_lib/auth'
-import { handleOptions, jsonError, methodNotAllowed, setCors, getStringQuery } from '../../../_lib/http'
-import { getProject, getRepoConfig, getShareById } from '../../../_lib/store'
-import { buildPrompt } from '../../../_lib/prompts'
-import { decryptToken } from '../../../_lib/tokens'
+import { requireReviewer } from '../../../../lib/auth'
+import { handleOptions, jsonError, methodNotAllowed, setCors, getStringQuery } from '../../../../lib/http'
+import { getProject, getRepoConfig, getShareById } from '../../../../lib/store'
+import { buildPrompt } from '../../../../lib/prompts'
+import { decryptToken } from '../../../../lib/tokens'
 
 export default async function handler(req: VercelRequest, res: VercelResponse) {
   if (handleOptions(req, res, ['GET', 'OPTIONS'])) return
