+++ lib/shares.ts
